Fix many compiler and cargo clippy warnings (#111)
* Suppress compiler warnings in generated flatbuffers code

There's no real need to surface these warnings since they originate from the
generated code and we have no way to fix them (without patching the flatbuffers
complier).

* Run cargo clippy --fix on epoch_publisher/

This fixes some, but not all of the compiler warnings. The remainder will need
to be addressed by hand.

* Convert <= check to == for epoch version

Since the epoch variable is unsigned, the less than check is unnecessary
and triggering a clippy warning.

* Run cargo clippy --fix in epoch/

This fixes some, but not all of the compiler warnings. The remainder will need
to be addressed by hand.

* Replace unnecessary if let with if status.is_err()

* Run cargo clippy --fix on common/

This fixes some, but not all of the compiler warnings. The remainder will need
to be addressed by hand.

* Fix clippy::inherent_to_string_shadow_display for HostPort

More details about why this is undesirable:
- https://rust-lang.github.io/rust-clippy/master/index.html#inherent_to_string_shadow_display

We're already implementing the Display trait for this type, so all we need to
do is delete the to_string() impl and rely on the one provided by the Display
trait.

* Run cargo clippy --fix on coordinator/

This fixes some, but not all of the compiler warnings. The remainder will need
to be addressed by hand.

* Replace unnecessary if let with option.is_some()

* Run cargo clippy --fix on universe/

This fixes some, but not all of the compiler warnings. The remainder will need
to be addressed by hand.

* Fix clippy::wrong_self_convention warning for to_keyspace_info

By convention this function should be renamed to `into_keyspace_info` since it
takes self by value instead of reference.

More info:

- https://rust-lang.github.io/rust-clippy/master/index.html#wrong_self_convention

* Fix clippy::unnecessary_map_or

More info:
  - https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_map_or